(review) Extract SuffixDialog, ConfirmDialog, FaceCard, useDropdownPosition (move-only)
Move four definitions out of ReviewModule.jsx (~1540 lines) into focused
files with byte-identical bodies (only relative import paths retargeted):

- components/review/SuffixDialog.jsx
- components/review/ConfirmDialog.jsx
- components/review/FaceCard.jsx
- hooks/useDropdownPosition.js (shared-hook convention)

ReviewModule imports them; the now-unused createPortal and useThumbnail
imports (only FaceCard used them) are dropped. No behavior change; all
characterization tests green without retargeting.
