+++ frontend/src/renderer/components/ReviewModule.jsx
@@ -10,7 +10,6 @@
  */
 
 import React, { useState, useEffect, useRef, useCallback } from 'react';
-import { createPortal } from 'react-dom';
 import { useModuleEvent, useEmitEvent } from '../hooks/useModuleEvent.js';
 import { useBackend } from '../context/BackendContext.jsx';
 import { useToast } from '../context/ToastContext.jsx';
@@ -18,71 +17,13 @@ import { useWebSocket } from '../hooks/useWebSocket.js';
 import { debug, debugWarn, debugError } from '../shared/debug.js';
 import { NetworkError } from '../shared/api-client.js';
 import { preferences } from '../workspace/preferences.js';
-import { useThumbnail } from '../shared/thumbnail-cache.js';
 import { normalizeSuffix } from '../shared/manualSuffix.js';
 import { Icon } from './Icon.jsx';
 import { t } from '../../i18n/index.js';
+import { SuffixDialog } from './review/SuffixDialog.jsx';
+import { ConfirmDialog } from './review/ConfirmDialog.jsx';
+import { FaceCard } from './review/FaceCard.jsx';
 import './ReviewModule.css';
-
-/**
- * Hook for calculating dropdown position with flip logic.
- * Renders dropdown below input if space available, otherwise above.
- * @param {boolean} open - Whether dropdown is open
- * @param {HTMLElement|null} anchorEl - The input element to anchor to
- * @param {Object} options - Configuration options
- * @returns {Object} Style object for the dropdown
- */
-function useDropdownPosition(open, anchorEl, { maxHeight = 200, gap = 4 } = {}) {
-  const [style, setStyle] = useState({ display: 'none' });
-
-  useEffect(() => {
-    if (!open || !anchorEl) {
-      setStyle({ display: 'none' });
-      return;
-    }
-
-    const updatePosition = () => {
-      const rect = anchorEl.getBoundingClientRect();
-      const viewportHeight = window.innerHeight;
-
-      const spaceBelow = viewportHeight - rect.bottom - gap;
-      const spaceAbove = rect.top - gap;
-
-      const placeAbove = spaceBelow < maxHeight && spaceAbove > spaceBelow;
-      const availableHeight = placeAbove ? Math.min(spaceAbove, maxHeight) : Math.min(spaceBelow, maxHeight);
-
-      const newStyle = {
-        position: 'fixed',
-        left: `${rect.left}px`,
-        width: `${rect.width}px`,
-        maxHeight: `${availableHeight}px`,
-        zIndex: 10001,
-      };
-
-      if (placeAbove) {
-        newStyle.bottom = `${viewportHeight - rect.top + gap}px`;
-        newStyle.top = 'auto';
-      } else {
-        newStyle.top = `${rect.bottom + gap}px`;
-        newStyle.bottom = 'auto';
-      }
-
-      setStyle(newStyle);
-    };
-
-    updatePosition();
-
-    window.addEventListener('scroll', updatePosition, true);
-    window.addEventListener('resize', updatePosition);
-
-    return () => {
-      window.removeEventListener('scroll', updatePosition, true);
-      window.removeEventListener('resize', updatePosition);
-    };
-  }, [open, anchorEl, maxHeight, gap]);
-
-  return style;
-}
 
 /**
  * ReviewModule Component
@@ -1175,362 +1116,6 @@ export function ReviewModule({ node }) {
           onCancel={() => setSuffixDialog(null)}
         />
       )}
-    </div>
-  );
-}
-
-/**
- * SuffixDialog Component
- *
- * Small dialog to attach a free-text filename suffix to the current image.
- * Shows a live, client-normalized preview mirroring the backend. Enter saves,
- * Esc cancels. Reuses the confirm-overlay/confirm-dialog styling.
- */
-function SuffixDialog({ initialValue, originalName, onSave, onCancel }) {
-  const [value, setValue] = useState(initialValue || '');
-  const inputRef = useRef(null);
-
-  useEffect(() => {
-    // Focus and select the field on open so re-editing is quick.
-    const el = inputRef.current;
-    if (el) {
-      el.focus();
-      el.select();
-    }
-  }, []);
-
-  const handleKeyDown = (e) => {
-    // Keep the dialog's own keys from reaching the module keyboard handler.
-    e.stopPropagation();
-    if (e.key === 'Enter') {
-      e.preventDefault();
-      onSave(value);
-    } else if (e.key === 'Escape') {
-      e.preventDefault();
-      onCancel();
-    }
-  };
-
-  const normalized = normalizeSuffix(value);
-  // Build a filename preview: strip an existing extension for display of the stem.
-  const dotIdx = originalName.lastIndexOf('.');
-  const stem = dotIdx > 0 ? originalName.slice(0, dotIdx) : originalName;
-  const ext = dotIdx > 0 ? originalName.slice(dotIdx) : '';
-  const previewName = normalized ? `${stem}_${normalized}${ext}` : `${stem}${ext}`;
-
-  return (
-    <div className="confirm-overlay" onClick={onCancel}>
-      <div
-        className="confirm-dialog"
-        tabIndex={-1}
-        onClick={(e) => e.stopPropagation()}
-      >
-        <h3>{t('review.manualSuffix.title')}</h3>
-        <input
-          ref={inputRef}
-          type="text"
-          className="suffix-input"
-          value={value}
-          placeholder={t('review.manualSuffix.placeholder')}
-          onChange={(e) => setValue(e.target.value)}
-          onKeyDown={handleKeyDown}
-        />
-        <div className="match-info">
-          {t('review.manualSuffix.previewLabel')} <strong>{previewName}</strong>
-        </div>
-        <div className="confirm-buttons">
-          <button className="btn-cancel" onClick={onCancel}>
-            {t('common.cancel')}
-          </button>
-          <button className="btn-confirm" onClick={() => onSave(value)}>
-            {t('common.save')}
-          </button>
-        </div>
-        <div className="confirm-hint">
-          {t('review.manualSuffix.hint')} · <kbd>Enter</kbd> {t('review.dialog.hintConfirms')} · <kbd>Esc</kbd> {t('review.dialog.hintCancels')}
-        </div>
-      </div>
-    </div>
-  );
-}
-
-function ConfirmDialog({ type, topMatch, chosenName, onConfirm, onCancel }) {
-  const dialogRef = useRef(null);
-
-  useEffect(() => {
-    const handleKeyDown = (e) => {
-      if (e.key === 'Enter') {
-        e.preventDefault();
-        onConfirm();
-      } else if (e.key === 'Escape') {
-        e.preventDefault();
-        onCancel();
-      }
-    };
-    document.addEventListener('keydown', handleKeyDown);
-    dialogRef.current?.focus();
-    return () => document.removeEventListener('keydown', handleKeyDown);
-  }, [onConfirm, onCancel]);
-
-  const isNameMismatch = type === 'name-mismatch';
-
-  return (
-    <div className="confirm-overlay" onClick={onCancel}>
-      <div
-        ref={dialogRef}
-        className="confirm-dialog"
-        tabIndex={-1}
-        onClick={(e) => e.stopPropagation()}
-      >
-        <h3>{isNameMismatch ? t('review.dialog.confirmNameChange') : t('review.dialog.confirmIgnore')}</h3>
-        <div className="match-info">
-          {t('review.dialog.bestMatch')} <strong>{topMatch.name}</strong> ({topMatch.confidence}%)
-        </div>
-        <p>
-          {isNameMismatch
-            ? t('review.dialog.nameMismatch', { name: chosenName })
-            : t('review.dialog.ignoreConfirm')}
-        </p>
-        <div className="confirm-buttons">
-          <button className="btn-cancel" onClick={onCancel}>
-            {t('common.cancel')}
-          </button>
-          <button className="btn-confirm" onClick={onConfirm}>
-            {t('common.confirm')}
-          </button>
-        </div>
-        <div className="confirm-hint">
-          <kbd>Enter</kbd> {t('review.dialog.hintConfirms')} · <kbd>Esc</kbd> {t('review.dialog.hintCancels')}
-        </div>
-      </div>
-    </div>
-  );
-}
-
-/**
- * FaceCard Component
- */
-function FaceCard({ face, index, isActive, imagePath, people, cardRef, inputRef, onSelect, onConfirm, onIgnore, onUnconfirm, maxAlternatives, onSelectAlternative, clearInputTrigger }) {
-  const isProbableIgnoreCase = face.match_case === 'ign' || face.match_case === 'uncertain_ign';
-  const initialValue = isProbableIgnoreCase ? '' : (face.person_name || '');
-  const [inputValue, setInputValue] = useState(initialValue);
-  const [typedValue, setTypedValue] = useState(initialValue);
-
-  // Use cached thumbnail
-  const { url: thumbnailUrl, loading: thumbnailLoading, error: thumbnailError } = useThumbnail(
-    imagePath,
-    face.bounding_box
-  );
-  const [showSuggestions, setShowSuggestions] = useState(false);
-  const [selectedSuggestion, setSelectedSuggestion] = useState(-1);
-  const localInputRef = useRef(null);
-
-  const setInputRef = useCallback((el) => {
-    localInputRef.current = el;
-    if (inputRef) {
-      if (typeof inputRef === 'function') inputRef(el);
-      else inputRef.current = el;
-    }
-  }, [inputRef]);
-
-  React.useEffect(() => {
-    const newValue = isProbableIgnoreCase ? '' : (face.person_name || '');
-    setInputValue(newValue);
-    setTypedValue(newValue);
-  }, [face.face_id, face.match_case, isProbableIgnoreCase, face.person_name]);
-
-  React.useEffect(() => {
-    if (clearInputTrigger > 0) {
-      setInputValue('');
-      setTypedValue('');
-    }
-  }, [clearInputTrigger]);
-
-  const filteredPeople = React.useMemo(() => {
-    if (!typedValue?.trim()) return [];
-    const typed = typedValue.toLowerCase();
-    const startsWithMatch = people.filter(p => p.toLowerCase().startsWith(typed));
-    const containsMatch = people.filter(p =>
-      !p.toLowerCase().startsWith(typed) && p.toLowerCase().includes(typed)
-    );
-    return [...startsWithMatch, ...containsMatch].slice(0, 8);
-  }, [typedValue, people]);
-
-  const dropdownStyle = useDropdownPosition(
-    showSuggestions && filteredPeople.length > 0,
-    localInputRef.current,
-    { maxHeight: 200, gap: 4 }
-  );
-
-  // Determine if this is a probable-ignore case
-  const isProbableIgnore = face.match_case === 'ign' || face.match_case === 'uncertain_ign';
-
-  const cardClass = [
-    'face-card',
-    face.is_confirmed && !face.is_rejected ? 'confirmed' : '',
-    face.is_rejected ? 'rejected' : '',
-    face.is_manual ? 'manual' : '',
-    isProbableIgnore && !face.is_confirmed ? 'probable-ignore' : '',
-    isActive ? 'active' : ''
-  ].filter(Boolean).join(' ');
-
-  const handleDoubleClick = (e) => {
-    if (face.is_confirmed && onUnconfirm) {
-      e.stopPropagation();
-      onUnconfirm();
-    }
-  };
-
-  return (
-    <div ref={cardRef} className={cardClass} onClick={onSelect} onDoubleClick={handleDoubleClick}>
-      <div className="face-number">{index + 1}</div>
-
-      <div className="face-thumbnail">
-        {thumbnailLoading ? (
-          <div className="thumbnail-loading" />
-        ) : thumbnailUrl && !thumbnailError ? (
-          <img
-            src={thumbnailUrl}
-            alt={face.person_name || t('review.unknown')}
-          />
-        ) : (
-          <Icon name="user" size={32} />
-        )}
-      </div>
-
-      <div className="face-info">
-        {/* Match case indicator */}
-        {face.is_manual && (
-          <div className="match-case manual">Manuellt tillagd</div>
-        )}
-        {face.match_case === 'ign' && !face.is_confirmed && (
-          <div className="match-case probable-ignore">Trolig ignorering</div>
-        )}
-        {face.match_case === 'uncertain_ign' && !face.is_confirmed && (
-          <div className="match-case uncertain">
-            ign ({face.ignore_confidence}%) / {face.person_name || face.match_alternatives?.[0]?.name || 'Okänd'}
-          </div>
-        )}
-        {face.match_case === 'uncertain_name' && !face.is_confirmed && (
-          <div className="match-case uncertain">
-            {face.person_name || face.match_alternatives?.[0]?.name || 'Okänd'} / ign ({face.ignore_confidence}%)
-          </div>
-        )}
-        {face.disambiguated && !face.is_confirmed && (
-          <div
-            className="match-case twin-disambig"
-            title={`Lika ansikten ${face.disambiguated.between.join(' / ')} — valt via k-NN-röstning över bekräftade foton`}
-          >
-            Tvilling-särskiljning → {face.disambiguated.chosen}
-          </div>
-        )}
-      </div>
-
-      {/* Match alternatives - only shown on active unconfirmed face */}
-      {isActive && !face.is_confirmed && face.match_alternatives?.length > 0 && (
-        <div className="face-alternatives">
-          {face.match_alternatives.slice(0, maxAlternatives || 5).map((alt, idx) => (
-            <div
-              key={idx}
-              className={`alt-chip ${idx === 0 ? 'recommended' : ''} ${alt.is_ignored ? 'ignored' : ''}`}
-              onClick={(e) => {
-                e.stopPropagation();
-                onSelectAlternative?.(alt.name);
-              }}
-            >
-              <span className="kbd">{idx + 1}</span>
-              <span className="alt-name">{alt.name}</span>
-              <span className="alt-conf">{alt.confidence}%</span>
-            </div>
-          ))}
-        </div>
-      )}
-
-      <div className="face-actions">
-        {!face.is_confirmed ? (
-          <div className="autocomplete-wrapper">
-            <input
-              ref={setInputRef}
-              type="text"
-              className={people.includes(inputValue) ? 'name-match' : ''}
-              placeholder={t('review.placeholder')}
-              value={inputValue}
-              onChange={(e) => {
-                const val = e.target.value;
-                setInputValue(val);
-                setTypedValue(val);
-                setSelectedSuggestion(-1);
-                setShowSuggestions(true);
-              }}
-              onFocus={() => setShowSuggestions(true)}
-              onBlur={() => setTimeout(() => setShowSuggestions(false), 150)}
-              onKeyDown={(e) => {
-                if (e.key === 'Escape') {
-                  setShowSuggestions(false);
-                  setInputValue(typedValue);
-                  setSelectedSuggestion(-1);
-                  e.target.blur();
-                  e.stopPropagation();
-                  return;
-                }
-                // Arrow Down / Tab - select next suggestion (wraps)
-                if ((e.key === 'ArrowDown' || (e.key === 'Tab' && !e.shiftKey)) &&
-                    showSuggestions && filteredPeople.length > 0) {
-                  e.preventDefault();
-                  const newIdx = selectedSuggestion >= filteredPeople.length - 1
-                    ? 0
-                    : selectedSuggestion + 1;
-                  setSelectedSuggestion(newIdx);
-                  setInputValue(filteredPeople[newIdx]);
-                  return;
-                }
-                // Arrow Up / Shift+Tab - select previous suggestion (wraps)
-                if ((e.key === 'ArrowUp' || (e.key === 'Tab' && e.shiftKey)) &&
-                    showSuggestions && filteredPeople.length > 0) {
-                  e.preventDefault();
-                  const newIdx = selectedSuggestion <= 0
-                    ? filteredPeople.length - 1
-                    : selectedSuggestion - 1;
-                  setSelectedSuggestion(newIdx);
-                  setInputValue(filteredPeople[newIdx]);
-                  return;
-                }
-              }}
-              onClick={(e) => e.stopPropagation()}
-            />
-            {showSuggestions && filteredPeople.length > 0 && createPortal(
-              <div className="autocomplete-dropdown" style={dropdownStyle}>
-                {filteredPeople.map((name, idx) => (
-                  <div
-                    key={name}
-                    className={`autocomplete-item ${idx === selectedSuggestion ? 'selected' : ''}`}
-                    onMouseDown={(e) => {
-                      e.preventDefault();
-                      setInputValue(name);
-                      setTypedValue(name);
-                      setShowSuggestions(false);
-                    }}
-                  >
-                    {name}
-                  </div>
-                ))}
-              </div>,
-              document.body
-            )}
-          </div>
-        ) : (
-          <div
-            className={`status-text ${face.is_rejected ? 'rejected' : 'confirmed'}`}
-            title={t('review.undoTitle')}
-          >
-            {face.is_rejected ? (
-              <><Icon name="block" size={12} /> {t('review.ignoredBadge')}</>
-            ) : (
-              <><Icon name="check" size={12} /> {face.person_name}</>
-            )}
-          </div>
-        )}
-      </div>
     </div>
   );
 }
